voice-chat-bot/requirements.txt
openai==1.12.0
python-dotenv==1.0.0
SpeechRecognition==3.10.0
sounddevice==0.4.6
numpy==1.24.3
flask==3.0.2 